Fixed signed/unsigned comparison warning in NetPacketNotify and NetPacketNotifyTests
#jira None
#rb Trivial

[CL 4656769 by Mattias Hornlund in Dev-Networking branch]

+++ Engine/Source/Runtime/Engine/Private/Net/Tests/NetPacketNotifyTests.cpp
@@ -49,7 +49,7 @@ struct FNetPacketNotifyTestUtil
 	{
 		size_t SequenceNumberCount = sizeof(InSequenceNumbers) / sizeof(InSequenceNumbers[0]);
 
-		for (int32 I=0; I<SequenceNumberCount; ++I)
+		for (size_t I=0; I<SequenceNumberCount; ++I)
 		{
 			FNetPacketNotifyTestUtil::PretendReceiveSeq(PacketNotify, InSequenceNumbers[I]);
 		}
